\name{sube-classes}
\title{sube workflow object classes}
\description{
The `sube` workflow returns a small set of object classes that represent each
major stage of the package API.
}
\details{
\describe{
  \item{\code{sube_suts}}{Imported supply-use data in normalized long form.}
  \item{\code{sube_domestic_suts}}{Domestic-block subset of imported SUT data.}
  \item{\code{sube_matrices}}{Aggregated supply/use matrices and final demand.}
  \item{\code{sube_results}}{Computed multipliers, elasticities, diagnostics,
  and coefficient matrices.}
  \item{\code{sube_models}}{OLS, pooled, between, and combined tidy model
  output tables.}
}
}
\keyword{classes}
