\name{sube-package}
\alias{sube-package}
\alias{sube}
\title{sube: Supply-Use Based Econometric Workflow Tools}
\description{
Reusable tools for importing supply-use tables, building domestic matrices,
computing SUBE multipliers, estimating regressions, and exporting tidy outputs.
}
\keyword{package}

